+++ src/main/java/org/eventticket/eventticketapp/EventDTO/request/EventUpdateRequest.java
@@ -1,22 +1,16 @@
-package org.eventticket.eventticketapp.dto.request;
+package org.eventticket.eventticketapp.EventDTO.request;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.time.LocalDateTime;
-
-@Setter
 @Getter
-public class BookEventRequest {
-
+@Setter
+public class EventUpdateRequest {
 
     private int eventId;
     private String eventName;
     private String eventDescription;
     private String eventLocation;
     private String eventDateAndTime;
-    private String price;
 
 }
