+++ src/main/java/org/eventticket/eventticketapp/EventDTO/request/FindEventByNameRequest.java
@@ -1,15 +1,15 @@
-package org.eventticket.eventticketapp.dto.response;
+package org.eventticket.eventticketapp.EventDTO.request;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class FindEventByLocationResponse {
+public class FindEventByNameRequest {
 
     private int eventId;
     private String eventName;
     private String location;
     private String description;
-
+    private String price;
 }
