+++ src/main/java/org/eventticket/eventticketapp/EventDTO/request/ViewAllEventsRequest.java
@@ -1,0 +1,6 @@
+package org.eventticket.eventticketapp.EventDTO.request;
+
+public class ViewAllEventsRequest {
+
+
+}
