+++ src/main/java/org/eventticket/eventticketapp/EventDTO/response/FindEventByCategoryResponse.java
@@ -1,16 +1,14 @@
-package org.eventticket.eventticketapp.dto.response;
+package org.eventticket.eventticketapp.EventDTO.response;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class FindEventByNameResponse {
-
+public class FindEventByCategoryResponse {
 
     private int eventId;
     private String eventName;
     private String location;
     private String description;
-    private String price;
 }
