+++ src/main/java/org/eventticket/eventticketapp/EventDTO/response/FindEventByLocationResponse.java
@@ -1,15 +1,15 @@
-package org.eventticket.eventticketapp.dto.request;
+package org.eventticket.eventticketapp.EventDTO.response;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class FindEventByNameRequest {
+public class FindEventByLocationResponse {
 
     private int eventId;
     private String eventName;
     private String location;
     private String description;
-    private String price;
+
 }
