+++ src/main/java/org/eventticket/eventticketapp/data/repository/EventRepository.java
@@ -1,10 +1,8 @@
 package org.eventticket.eventticketapp.data.repository;
 
 import org.eventticket.eventticketapp.data.model.Event;
-import org.eventticket.eventticketapp.dto.response.FindEventByPriceResponse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
 import java.util.Optional;
 
 public interface EventRepository extends JpaRepository<Event, Integer> {
