--- src/main/java/org/eventticket/eventticketapp/dto/request/ViewAllEventsRequest.java
@@ -1,6 +1,0 @@
-package org.eventticket.eventticketapp.dto.request;
-
-public class ViewAllEventsRequest {
-
-
-}
