+++ src/main/java/org/eventticket/eventticketapp/services/AdminServices.java
@@ -1,8 +1,8 @@
 package org.eventticket.eventticketapp.services;
 
 import org.eventticket.eventticketapp.data.model.Event;
-import org.eventticket.eventticketapp.dto.request.*;
-import org.eventticket.eventticketapp.dto.response.*;
+import org.eventticket.eventticketapp.EventDTO.request.*;
+import org.eventticket.eventticketapp.EventDTO.response.*;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
