+++ src/test/java/org/eventticket/eventticketapp/EventTicketAppApplicationTests.java
@@ -1,19 +1,16 @@
 package org.eventticket.eventticketapp;
 
-import org.eventticket.eventticketapp.data.model.Admin;
 import org.eventticket.eventticketapp.data.model.Event;
 import org.eventticket.eventticketapp.data.repository.EventRepository;
-import org.eventticket.eventticketapp.dto.request.*;
-import org.eventticket.eventticketapp.dto.response.*;
+import org.eventticket.eventticketapp.EventDTO.request.*;
+import org.eventticket.eventticketapp.EventDTO.response.*;
 import org.eventticket.eventticketapp.services.AdminServices;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+
 import org.eventticket.eventticketapp.adminServices.AdminService;
-import org.eventticket.eventticketapp.data.repository.AdminRepository;
 import org.eventticket.eventticketapp.dtos.request.CreateLoginRequest;
 import org.eventticket.eventticketapp.dtos.request.CreateLogoutRequest;
 import org.eventticket.eventticketapp.dtos.request.CreateRegistrationRequest;
@@ -24,8 +21,6 @@ import org.eventticket.eventticketapp.exception.UserAlreadyExistException;
 
 import java.util.List;
 import java.util.Optional;
-
-import static org.assertj.core.api.Assertions.*;
 
 
 @SpringBootTest
